Separated GameController methods into GameController, LobbyController and MatchController
