+++ src/main/java/org/springframework/samples/dobble/game/GameController.java
@@ -94,12 +94,12 @@ public class GameController {
 
     @PostMapping("/new")
     public String createGame(Game game, BindingResult result) {
+        if (result.hasErrors())
+            return VIEWS_GAMES_CREATE_OR_UPDATE_FORM;
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         String userId = authentication.getName();
         User owner = userService.findUser(userId);
         game.setOwner(owner);
-        if (result.hasErrors())
-            return VIEWS_GAMES_CREATE_OR_UPDATE_FORM;
         game.setState(GameState.LOBBY);
         this.gameService.saveGame(game);
         return "redirect:/games/" + game.getId();
@@ -119,17 +119,7 @@ public class GameController {
         return "redirect:/games/{gameId}/lobby";
 
     }
-  
-    @GetMapping(path = "/{gameId}/play/delete/{userId}")
-    public String deleteUsersGame(@PathVariable("gameId") Long gameId,
-            @PathVariable("userId") String userId, RedirectAttributes redirAttrs) {
-        try {
-            gameService.deleteGameUser(gameId, userId);
-        } catch (Exception e) {
-            return "redirect:/games?error=" + e.getMessage();
-        }
-        return "redirect:/games/{gameId}/play";
-    }
+
 
     @GetMapping("/{gameId}/start")
     public String startGame(@PathVariable("gameId") Long gameId) {
@@ -152,57 +142,10 @@ public class GameController {
 
     // In-game related actions
 
-    @GetMapping("/{gameId}/play")
-    public ModelAndView playGame(@PathVariable("gameId") Long gameId) {
-        ModelAndView mav = new ModelAndView(VIEW_PLAY_GAME);
-        Game game = this.gameService.findGame(gameId);
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String username = authentication.getName();
-        User mainUser = userService.findUser(username);
-        GameUser mainPlayer = gameUserService.findGameUser(GameUserPk.of(mainUser , game));
-        List<GameUser> players = game.getUsers();
-        players.removeIf(player -> player.equals(mainPlayer));
-        mav.addObject("mainPlayer", mainPlayer);
-        mav.addObject("players", players);
-        mav.addObject("game", game);
-        mav.addObject(game);
-        return mav;
+   
+  
 
-    }
-    @GetMapping("/{gameId}/lobby")
-    public ModelAndView lobbyGame(@PathVariable("gameId") Long gameId) {
-        Game game = this.gameService.findGame(gameId);
-        Iterable<GameUser> mazos=game.getUsers();
-		ModelAndView result=new ModelAndView("games/LobbyGame");
-		result.addObject("users", mazos);
-        result.addObject("game", game);
-		return result;	
-    }
-
-
-    @PostMapping("/{gameId}/match")
-    public String checkMatch(@PathVariable("gameId") Long gameId, @ModelAttribute("symbol") Long symbolId) {
-        //This method steps are only made for testing at the moment. For the next sprint it will 
-        //fully implement the required mehtod
-        System.out.println("MATCH");
-        Game game = gameService.findGame(gameId);
-        Symbol symbol = symbolService.findById(symbolId);
-        Boolean test = game.getCurrentCard().getSymbols().contains(symbol);
-        if (test) game.nextCard();;
-        gameService.saveGame(game);
-        if (game.getCentralDeck().size()==0) return "redirect:play?NoMoreCardsInTheCenter";
-        return "redirect:play?" + test;
-    }
-
-    @GetMapping(path="/{gameId}/lobby/delete/{id}")
-	public String DeleteUsersGame(@PathVariable("gameId") Long gameId, @PathVariable("id") String id, RedirectAttributes redirAttrs) {
-       try {
-            gameService.deleteGameUser(gameId, id);
-       } catch(Exception e) {
-           return "redirect:/games?error="+ e.getMessage();
-        } 
-        return "redirect:/games/{gameId}/lobby";
-    }
+    
 
 
 }
